# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/budget": {
            "post": {
                "description": "Creates a new budget for the user with the provided details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Create a new budget",
                "operationId": "create-budget",
                "parameters": [
                    {
                        "description": "Budget Details",
                        "name": "budget",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/finly-backend_internal_service_budget.CreateBudgetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/finly-backend_internal_service_budget.CreateBudgetResponse"
                        }
                    }
                }
            }
        },
        "/budget/{budget_id}": {
            "get": {
                "description": "Retrieves the current balance of a budget for the specified user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Get current balance",
                "operationId": "get-current-balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "budget_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/finly-backend_internal_service_budget.GetCurrentBalanceResponse"
                        }
                    }
                }
            }
        },
        "/budget/{budget_id}/history": {
            "get": {
                "description": "Retrieves the history of a budget for the specified user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Get budget history",
                "operationId": "get-budget-history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Budget ID",
                        "name": "budget_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/finly-backend_internal_service_budget.GetBudgetHistoryResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "finly-backend_internal_service_budget.BudgetHistory": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "budget_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "finly-backend_internal_service_budget.CreateBudgetRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "userID"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "finly-backend_internal_service_budget.CreateBudgetResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "finly-backend_internal_service_budget.GetBudgetHistoryResponse": {
            "type": "object",
            "properties": {
                "budget_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/finly-backend_internal_service_budget.BudgetHistory"
                    }
                }
            }
        },
        "finly-backend_internal_service_budget.GetCurrentBalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        },
...
    }
}